delete Core component
